Add TMB sdreport object to WHAM model

Runs TMB::sdreport and adds the object to the fitted (and projected) model list. E.g., fit$sdrep.

Usage

do_sdreport(
  model,
  save.sdrep = TRUE,
  TMB.bias.correct = FALSE,
  TMB.jointPrecision = FALSE
)

Arguments

model

a fitted WHAM model object returned by fit_wham or project_wham.

save.sdrep

T/F, save the full TMB::sdreport object? If FALSE, only save summary.sdreport to reduce model object file size. Default = TRUE.

TMB.bias.correct

T/F whether to use the bias.correct feature of TMB::sdreport. Default = FALSE.

TMB.jointPrecision

T/F whether to return the joint precision matrix for the fixed and random effects from TMB::sdreport. Default = FALSE.

See also