Prepare input data and parameters to add projection to a WHAM operating model

prepare_wham_om_proj is an internal function called by prepare_wham_om_input if do.proj = TRUE.

prepare_wham_om_proj(om_input, proj.opts)

Arguments

om_input
is output from prepare_wham_om_input with do.proj = FALSE

proj.opts
a named list with the following components:

$n.yrs (integer), number of years to project/forecast. Default = 3.

$use.last.F (T/F), use terminal year F for projections. Default = TRUE.

$use.FXSPR (T/F), calculate and use F at X

$use.FMSY (T/F), calculate and use FMSY for projections.

$proj.F (vector), user-specified fishing mortality for projections. Length must equal n.yrs.

$proj.catch (vector), user-specified aggregate catch for projections. Length must equal n.yrs.

$avg.yrs (vector), specify which years to average over for calculating reference points. Default = last 5 model years, tail(model$years, 5).

$cont.ecov (T/F), continue ecov process (e.g. random walk or AR1) for projections. Default = TRUE.

$use.last.ecov (T/F), use terminal year ecov for projections.

$avg.ecov.yrs (vector), specify which years to average over the environmental covariate(s) for projections.

$proj.ecov (vector), user-specified environmental covariate(s) for projections. Length must equal n.yrs.

$cont.Mre (T/F), continue M random effects (i.e. AR1_y or 2D AR1) for projections. Default = TRUE. If FALSE, M will be averaged over $avg.yrs (which defaults to last 5 model years).

$avg.rec.yrs (vector), specify which years to calculate the CDF of recruitment for use in projections. Default = all model years.

$percentFXSPR (scalar), percent of F_XSPR to use for calculating catch in projections, only used if $use.FXSPR = TRUE. For example, GOM cod uses F = 75

$percentFMSY (scalar), percent of F_MSY to use for calculating catch in projections, only used if $use.FMSY = TRUE.

Value

same as prepare_wham_om_input, a list ready to use with TMB::MakeADFun to simulate populations and/or data

data

Named list of data, passed to TMB::MakeADFun

par

Named list of parameters, passed to TMB::MakeADFun

map

Named list of factors that determine which parameters are estimated, passed to TMB::MakeADFun

random

Character vector of parameters to treat as random effects, passed to TMB::MakeADFun

years

Numeric vector of representing (non-projection) model years of WHAM model

years_full

Numeric vector of representing all model and projection years of WHAM model

ages.lab

Character vector of age labels, ending with plus-group

model_name

Character, name of stock/model (specified in call to prepare_wham_input)

See also