Update input data and parameters from simulated data with projection

When data are simulated from a wham model with projections, this function can be used to update the input model years. This is helpful for using WHAM as an assessment model in a management strategy evaluation. See prepare_wham_input, prepare_wham_om_input, prepare_wham_om_proj, and project_wham for further details on data and parameter structure.

update_wham_input(simres, model, n_years_add)

Arguments

simres
list containing data and parameters (output from model$simulate(complete=TRUE), where model is output from fit_wham)

model
object used to produce simres.

n_years_add
number of projection years to convert to model_years

Value

a named list with the following components:

data

Named list of data, passed to TMB::MakeADFun

par

Named list of parameters, passed to TMB::MakeADFun

map

Named list defining how to optionally collect and fix parameters, passed to TMB::MakeADFun

random

Character vector of parameters to treat as random effects, passed to TMB::MakeADFun

years

Numeric vector of years to fit WHAM model

years_full

Numeric vector of years to fit and project WHAM model

ages.lab

Character vector of age labels, ending with plus-group

model_name

Character, name of stock/model (specified in call to prepare_wham_input)

See also

Examples

if (FALSE) {
input = prepare_wham_om_input(asap3)
mod = fit_wham(input, do.fit = FALSE, do.proj = TRUE)
sim = mod$simulate(complete=TRUE)
sim_input = update_wham_input(sim, mod, 3)
simfit = fit_wham(sim_input, MakeADFun.silent = TRUE)
plot_wham_output(simfit)
}