Fit WHAM model

Fits the compiled WHAM model using TMB::MakeADFun and stats::nlminb. Runs retrospective analysis if specified.

fit_wham(
  input,
  n.newton = 3,
  do.sdrep = TRUE,
  do.retro = TRUE,
  n.peels = 7,
  do.osa = TRUE,
  osa.opts = list(method = "oneStepGeneric", parallel = TRUE),
  model = NULL,
  do.check = FALSE,
  do.proj = FALSE,
  proj.opts = list(n.yrs = 3, use.last.F = TRUE, use.avg.F = FALSE, use.FXSPR = FALSE,
    proj.F = NULL, proj.catch = NULL, avg.yrs = NULL, cont.Ecov = TRUE, use.last.Ecov =
    FALSE, avg.Ecov.yrs = NULL, proj.Ecov = NULL)
)

Arguments

input
Named list with components:

$data

Data to fit the assessment model to.

$par

Parameters, a list of all parameter objects required by the user template (both random and fixed effects). See MakeADFun.

$map

Map, a mechanism for collecting and fixing parameters. See MakeADFun.

$random

Character vector defining the parameters to treat as random effect. See MakeADFun.

$years

Numeric vector of the years which the model spans. Not important for model fitting, but useful for plotting.

$model_name

Character, name of the model, e.g. "Yellowtail flounder"

$ages.lab

Character vector of the age labels, e.g. c("1","2","3","4+").

n.newton
integer, number of additional Newton steps after optimization. Passed to fit_tmb. Default = 3.

do.sdrep
T/F, calculate standard deviations of model parameters? See sdreport. Default = TRUE.

do.retro
T/F, do retrospective analysis? Default = TRUE.

n.peels
integer, number of peels to use in retrospective analysis. Default = 7.

do.osa
T/F, calculate one-step-ahead (OSA) residuals? Default = TRUE. See details. Returned as mod$osa$residual.

osa.opts
list of options for calculating OSA residuals, passed to TMB::oneStepPredict. Default: osa.opts = list(method="oneStepGeneric", parallel=TRUE).

model
(optional), a previously fit wham model.

do.check
T/F, check if model parameters are identifiable? Passed to fit_tmb. Runs TMBhelper::Check_Identifiable. Default = TRUE.

do.proj
T/F, do projections? Default = FALSE. If true, runs project_wham.

proj.opts
list of options for projections

$n.yrs (integer), number of years to project/forecast. Default = 3.

$use.last.F (T/F), use terminal year F for projections. Default = TRUE.

$use.FXSPR (T/F), calculate F at X

$proj.F (vector), user-specified fishing mortality for projections. Length must equal n.yrs.

$proj.catch (vector), user-specified aggregate catch for projections. Length must equal n.yrs.

$avg.yrs (vector), specify which years to average over for calculating reference points. Default = last 5 model years, tail(model$years, 5).

$cont.Ecov (T/F), continue Ecov process (e.g. random walk or AR1) for projections. Default = TRUE.

$use.last.Ecov (T/F), use terminal year Ecov for projections.

$avg.Ecov.yrs (vector), specify which years to average over the environmental covariate(s) for projections.

$proj.Ecov (vector), user-specified environmental covariate(s) for projections. Length must equal n.yrs.

Value

a fit TMB model with additional output if specified:

$rep

List of derived quantity estimates (see examples)

$sdrep

Parameter estimates (and standard errors if do.sdrep=TRUE)

$peels

Retrospective analysis (if do.retro=TRUE)

$osa

One-step-ahead residuals (if do.osa=TRUE)

Details

Standard residuals are not appropriate for models with random effects. Instead, one-step-ahead (OSA) residuals can be used for evaluating model goodness-of-fit (Thygeson et al. (2017), implemented in TMB::oneStepPredict). Additional OSA residual options are passed to TMB::oneStepPredict in a list osa.opts. For example, to use the (much faster, ~1 sec instead of 2 min) full Gaussian approximation instead of the (default) generic method, you can use osa.opts=list(method="fullGaussian").

See also

Examples

if (FALSE) {
data("input4_SNEMAYT") # load SNEMA yellowtail flounder data and parameter settings
mod = fit_wham(input4_SNEMAYT) # using default values
mod = fit_wham(input4_SNEMAYT, do.retro=FALSE, osa.opts=list(method="fullGaussian")) # faster settings for initial model fitting

names(mod$rep) # list of derived quantities
mod$rep$SSB # get SSB estimates (weight, not numbers)
m1$rep$NAA[,1] # get recruitment estimates (numbers, first column of numbers-at-age matrix)
m1$rep$F[,1] # get F estimates for fleet 1
}