Prepare input data and parameters for WHAM model

After the data file is read into R by read_asap3_dat, this function prepares the data and parameter settings for fit_wham. By default, this will set up a SCAA version like ASAP.

prepare_wham_input(
  asap3,
  model_name = "WHAM for unnamed stock",
  recruit_model = 2,
  ecov = NULL,
  selectivity = NULL,
  M = NULL,
  NAA_re = NULL
)

Arguments

asap3
list containing data and parameters (output from read_asap3_dat)

model_name
character, name of stock/model

recruit_model
numeric, option to specify stock-recruit model (see details)

ecov
(optional) named list of environmental covariate data and parameters (see details)

selectivity
(optional) list specifying selectivity options by block: models, initial values, parameters to fix, and random effects (see details)

M
(optional) list specifying natural mortality options: model, random effects, initial values, and parameters to fix (see details)

NAA_re
(optional) list specifying random effect options on numbers-at-age (see details)

Value

a named list with the following components:

data

Named list of data, passed to TMB::MakeADFun

par

Named list of parameters, passed to TMB::MakeADFun

map

Named list defining how to optionally collect and fix parameters, passed to TMB::MakeADFun

random

Character vector of parameters to treat as random effects, passed to TMB::MakeADFun

years

Numeric vector of years to fit WHAM model (specified in ASAP3 .dat file)

ages.lab

Character vector of age labels, ending with plus-group (specified in ASAP3 .dat file)

model_name

Character, name of stock/model (specified in call to prepare_wham_input)

Details

recruit_model specifies the stock-recruit model. See wham.cpp for implementation.

= 1

Random walk, i.e. predicted recruitment in year i = recruitment in year i-1

= 2

(default) Random about mean, i.e. steepness = 1

= 3

Beverton-Holt

= 4

Ricker

Note: we allow fitting a SCAA model (NAA_re = NULL), which estimates recruitment in every year as separate fixed effect parameters, but in that case no stock-recruit function is estimated. A warning message is printed if recruit_model > 2 and NAA_re = NULL. If you wish to use a stock-recruit function for expected recruitment, choose recruitment deviations as random effects, either only age-1 (NAA_re = list(sigma='rec')) or all ages (NAA_re = list(sigma='rec+1'), "full state-space" model). See below for details on NAA_re specification.

ecov specifies any environmental covariate data and model. Environmental covariate data need not span the same years as the fisheries data. It can be NULL if no environmental data are to be fit. Otherwise, it must be a named list with the following components:

$label

Name(s) of the environmental covariate(s). Used in printing.

$mean

Mean observations (matrix). Missing values = NA.

$logsigma

Observation standard error (log). Options:
Matrix with same dimensions as $mean

Specified values (not estimated) for each time step

Single value per ecov, numeric vector or matrix w/ dim 1 x n.ecov

Specified value (not estimated) shared among time steps

Vector with estimate options for each ecov, length = n.ecov

'est_1': Estimated, one value shared among time steps. 'est_re': Estimated value for each time step as random effects with two parameters (mean, var)

$year

Years corresponding to observations (vector of same length as $mean and $logsigma)

$use_obs

T/F (or 0/1) vector/matrix of the same dimension as $mean and $logsigma. Use the observation? Can be used to ignore subsets of the ecov without changing data files.

$lag

Offset between the ecov observations and their affect on the stock. I.e. if SST in year t affects recruitment in year t + 1, set lag = 1.

$process_model

Process model for the ecov time-series. "rw" = random walk, "ar1" = 1st order autoregressive, NA = do not fit

$where

Where does the ecov affect the population? "recuit" = recruitment, "M" = natural mortality, "growth" = growth.

$how

How does the ecov affect the $where process? These options are specific to the $where process.

$link_model

Model describing ecov effect on the $where process. Options: 'linear' (default) or 'poly-x' where x = 2, ... (e.g. 'poly-2' specifies a quadratic model, \(b0 + b1*ecov + b2*ecov^2 + ...\)).

ecov$how specifies HOW the ecov affects the ecov$where process.

= 0

none (but fit ecov time-series model in order to compare AIC)

= 1

"controlling" (dens-indep mortality)

= 2

"limiting" (carrying capacity, e.g. ecov determines amount of suitable habitat)

= 3

"lethal" (threshold, i.e. R --> 0 at some ecov value)

= 4

"masking" (metabolic/growth, decreases dR/dS)

= 5

"directive" (e.g. behavioral)

Options for M:

= 0

none (but fit ecov time-series model in order to compare AIC)

= 1

effect on mean M (shared across ages)

selectivity specifies options for selectivity, to overwrite existing options specified in the ASAP data file. If NULL, selectivity options from the ASAP data file are used. selectivity is a list with the following entries:

$model

Selectivity model for each block. Vector with length = number of selectivity blocks. Each entry must be one of: "age-specific", "logistic", "double-logistic", or "decreasing-logistic".

$re

Time-varying (random effects) for each block. Vector with length = number of selectivity blocks. If NULL, selectivity parameters in all blocks are constant over time and uncorrelated. Each entry of selectivity$re must be one of the following options, where the selectivity parameters are:
"none"

(default) are constant and uncorrelated

"iid"

vary by year and age/par, but uncorrelated

"ar1"

correlated by age/par (AR1), but not year

"ar1_y"

correlated by year (AR1), but not age/par

"2dar1"

correlated by year and age/par (2D AR1)

$initial_pars

Initial parameter values for each block. List of length = number of selectivity blocks. Each entry must be a vector of length # parameters in the block, i.e. c(2,0.2) for logistic or c(0.5,0.5,0.5,1,1,0.5) for age-specific with 6 ages.

$fix_pars

Which parameters to fix at initial values. List of length = number of selectivity blocks. E.g. model with 3 age-specific blocks and 6 ages, list(c(4,5),4,c(2,3,4)) will fix ages 4 and 5 in block 1, age 4 in block 2, and ages 2, 3, and 4 in block 3.

M specifies estimation options for natural mortality and can overwrite M-at-age values specified in the ASAP data file. If NULL, the M-at-age matrix from the ASAP data file is used (M fixed, not estimated). To estimate M-at-age shared/mirrored among some but not all ages, modify input$map$M_a after calling prepare_wham_input (see vignette for more details). M is a list with the following entries:

$model

Natural mortality model options are:
"constant"

(default) estimate a single mean M shared across all ages

"age-specific"

estimate M_a independent for each age

"weight-at-age"

specifies M as a function of weight-at-age, \(M_y,a = exp(b0 + b1*log(W_y,a))\), as in Lorenzen (1996) and Miller & Hyun (2018).

$re

Time- and age-varying (random effects) on M. Note that random effects can only be estimated if mean M-at-age parameters are ($est_ages is not NULL).
"none"

(default) M constant in time and across ages.

"iid"

M varies by year and age, but uncorrelated.

"ar1_a"

M correlated by age (AR1), constant in time.

"ar1_y"

M correlated by year (AR1), constant all ages.

"2dar1"

M correlated by year and age (2D AR1), as in Cadigan (2016).

$initial_means

Initial/mean M-at-age vector, with length = number of ages (if $model = "age-specific") or 1 (if $model = "weight-at-age" or "constant"). If NULL, initial mean M-at-age values are taken from the first row of the MAA matrix in the ASAP data file.

$est_ages

Vector of ages to estimate M (others will be fixed at initial values). E.g. in a model with 6 ages, $est_ages = 5:6 will estimate M for the 5th and 6th ages, and fix M for ages 1-4. If NULL, M at all ages is fixed at M$initial_means (if not NULL) or row 1 of the MAA matrix from the ASAP file (if M$initial_means = NULL).

$logb_prior

(Only if $model = "weight-at-age") TRUE or FALSE (default), should a N(0.305, 0.08) prior be used on log_b? Based on Fig. 1 and Table 1 (marine fish) in Lorenzen (1996).

NAA_re specifies options for random effects on numbers-at-age (NAA, i.e. state-space model or not) If NULL, a traditional statistical catch-at-age model is fit (NAA = pred_NAA for all ages, deterministic). To fit a state-space model, specify NAA_re, a list with the following entries:

$sigma

Which ages allow deviations from pred_NAA? Common options are specified with the strings:
"rec"

Random effects on recruitment (deviations), all other ages deterministic

"rec+1"

"Full state space" model with 2 estimated sigma_a, one for recruitment and one shared among other ages

Alternatively, you can specify a more complex structure by entering a vector with length = n.ages, where each entry points to the NAA_sigma to use for that age. E.g. c(1,2,2,3,3,3) will estimate 3 sigma_a, with recruitment (age-1) deviations having their own sigma_R, ages 2-3 sharing sigma_2, and ages 4-6 sharing sigma_3.

$cor

Correlation structure for the NAA deviations. Options are:
"iid"

NAA deviations vary by year and age, but uncorrelated.

"ar1_a"

NAA deviations correlated by age (AR1).

"ar1_y"

NAA deviations correlated by year (AR1).

"2dar1"

NAA deviations correlated by year and age (2D AR1).

See also

Examples

if (FALSE) {
asap3 = read_asap3_dat("ASAP_SNEMAYT.dat")
input = prepare_wham_input(asap3)
mod = fit_wham(input)
}