Specify model and parameter configuration for numbers at age

Specify model and parameter configuration for numbers at age

set_NAA(input, NAA_re = NULL)

Arguments

input

list containing data, parameters, map, and random elements (output from prepare_wham_input)

NAA_re

(optional) list specifying options for numbers-at-age random effects, initial parameter values, and recruitment model (see details)

If NAA_re = NULL, a traditional statistical catch-at-age model is fit (NAA = pred_NAA for all ages, deterministic). Otherwise, NAA_re specifies numbers-at-age configuration. It is a list with the following possible entries:

$sigma

Which ages allow deviations from the predicted NAA given NAA from previou time step? Must be a single character string described below or a vector of length n_stocks. If length = 1, assumptions will be applied to all stocks. Common options are specified with the strings:

"rec"

Random effects on recruitment (deviations), all other ages deterministic

"rec+1"

"Full state space" model with 2 estimated sigma_a, one for recruitment and one shared among other ages

$sigma_map

You can specify a more complex parameter configuration by entering an integer array (nstocks x n_regions x n_ages), where each entry points to the NAA_sigma to use for that stock and age. E.g., for 2 stocks, 2 regions, and 6 ages, array(rep(c(1,2,2,3,3,3), each = 4),c(2,2,6)) will estimate 3 sigmas, with recruitment (age-1) deviations having their own sigma_R, ages 2-3 sharing sigma_2, and ages 4-6 sharing sigma_3. All parameters being the same for both stocks and across regions. The user must be sure that a compatible NAA_re$sigma configuration is defined. Values are not used if recruit_model = 1 and NAA_re$sigma is not specified.

$sigma_vals

Initial standard deviation values to use for the NAA deviations. Values are not used if recruit_model = 1 and NAA_re$sigma is not specifed. Otherwise when NAA_re$sigma =

"rec"

must be a list (length = n_stocks) of single values .

"rec+1"

Either 1) a list (length = n_stocks) of 2 values must be specified. First is for the first age class (recruits), second is for all other ages, or 2) an array (nstocks x nregions x nages)

only the value for age 1 in the spawning region is used.

If NAA_re$sigma_map is defined, the user must ensure that the configuration is compatible with NAA_re$sigma_vals

$cor

Correlation structure for the NAA deviations. Options are:

"iid"

NAA deviations vary by year and age, but uncorrelated.

"ar1_a"

NAA deviations correlated by age (AR1).

"ar1_y"

NAA deviations correlated by year (AR1).

"2dar1"

NAA deviations correlated by year and age (2D AR1).

$cor_vals

Initial correlation values to use for the NAA deviations. If unspecified all initial values are 0. When NAA_re$cor =

"iid"

values are not used.

"ar1_a" or "ar1_y"

cor_vals must be a single value.

"2dar1"

2 values must be specified. First is for "age", second is for "year".

$N1_model

Integer vector (n_stocks) determining which way to model the initial numbers at age:

"age-specific-fe"

(default) age- and region-specific fixed effects parameters

"equilibrium"

2 fixed effects parameters: an initial recruitment and an instantaneous fishing mortality rate to generate an equilibruim abundance at age.

"iid-re"

(default) age- and region-specific iid random effects parameters. 2 parameters: mean and sd for log NAA

"ar1-re"

(default) age- and region-specific random effects parameters. 3 parameters: mean and sd, and cor for log NAA

$N1_pars

if N1_model = 0, then these would be the initial values to use for abundance at age in the first year. If N1_model = 1, This would be the initial numbers in the first age class and the equilibrium fishing mortality rate generating the rest of the numbers at age in the first year.

$recruit_model

Integer vector (n_stocks) determining how to model recruitment. Overrides recruit_model argument to prepare_wham_input. Must make sure NAA_re$sigma, NAA_re$cor and ecov are properly specified.

1

SCAA, estimating all recruitements as fixed effects or a random walk if NAA_re$sigma specified

2

estimating a mean recruitment with yearly recruitements as random effects

3

Beverton-Holt stock-recruitment with yearly recruitements as random effects

4

Ricker stock-recruitment with yearly recruitements as random effects

$recruit_pars

list (length = n_stocks) of vectors of initial parameters for recruitment model. If use_steepness=F, parameters are "alpha" and "beta" otherwise they are steepness and R0.